I've figured it out
A lot of this is wrong. But what is supposed to happen is the
ObjectUpdate comes in with very basic params. This contains just the
UUID of the agent it's trying to render. Then it receives following
ObjectUpdate packets which contain AttachmentItem parameters. These are
applied to the empty shell it just asked you to spawn. These contain all
of the wearables attached to it. This makes it so that users MUST render
in semi-nude, since there is no way of knowing when the ObjectUpdate
agent attach packets will be done coming in. Oh my god.